rag-production/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
torch
numpy